refactor(cost): extract average cost pool repository
