+++ src/services/portfolio_transaction_processing_service/app/infrastructure/cost_repository.py
@@ -5,7 +5,6 @@ from typing import Any
 
 from portfolio_common.database_models import (
     AccruedIncomeOffsetState,
-    AverageCostPoolState,
     PositionLotState,
     TransactionCost,
 )
@@ -15,7 +14,7 @@ from portfolio_common.database_models import (
 from portfolio_common.domain.currency import normalize_currency_code
 from portfolio_common.events import TransactionEvent, event_business_payload
 from portfolio_common.identifiers import normalize_lookup_identifier
-from sqlalchemy import func, select, update
+from sqlalchemy import func, select
 from sqlalchemy.dialects.postgresql import insert as pg_insert
 from sqlalchemy.ext.asyncio import AsyncSession
 
@@ -35,6 +34,7 @@ from ..ports import (
     OpenLotCheckpointRecord,
 )
 from .booked_transaction_event_mapper import to_booked_transaction
+from .cost_basis.average_cost_pool_repository import SqlAlchemyAverageCostPoolRepository
 from .cost_basis.lot_state_mapper import buy_lot_state_payload, mutable_lot_state_fields
 
 TRANSACTION_METADATA_FIELDS = (
@@ -167,28 +167,10 @@ def _transaction_cost_rows(
     ]
 
 
-def _scaled_persisted_value(
-    column: Any,
-    *,
-    before: Decimal,
-    after: Decimal,
-    round_down: bool,
-) -> Any:
-    if after == before:
-        return column
-    if after == Decimal(0):
-        return Decimal(0)
-    if before <= Decimal(0):
-        raise ValueError("Average cost source scaling requires a positive prior aggregate")
-    scaled = column * after / before
-    return func.trunc(scaled, 10) if round_down else func.round(scaled, 10)
-
-
 class CostCalculatorRepository:
-    AVERAGE_COST_REBUILD_UPSERT_CHUNK_SIZE = 500
-
     def __init__(self, db: AsyncSession):
         self.db = db
+        self._average_cost_pools = SqlAlchemyAverageCostPoolRepository(db)
 
     async def get_transaction_history(
         self, portfolio_id: str, security_id: str, exclude_id: str | None = None
@@ -225,128 +207,28 @@ class CostCalculatorRepository:
         portfolio_id: str,
         security_id: str,
     ) -> AverageCostPoolCheckpointRecord | None:
-        stmt = (
-            select(AverageCostPoolState, DBTransaction)
-            .outerjoin(
-                DBTransaction,
-                DBTransaction.transaction_id
-                == AverageCostPoolState.representative_source_transaction_id,
-            )
-            .where(
-                AverageCostPoolState.portfolio_id == normalize_lookup_identifier(portfolio_id),
-                AverageCostPoolState.security_id == normalize_lookup_identifier(security_id),
-            )
-            .with_for_update(of=AverageCostPoolState)
-        )
-        row = (await self.db.execute(stmt)).first()
-        if row is None:
-            return None
-        state, representative_transaction = row
-        return AverageCostPoolCheckpointRecord(
-            checkpoint=AverageCostPoolCheckpoint(
-                portfolio_id=state.portfolio_id,
-                instrument_id=state.instrument_id,
-                security_id=state.security_id,
-                representative_source_transaction_id=(state.representative_source_transaction_id),
-                quantity=state.pool_quantity,
-                cost_local=state.pool_cost_local,
-                cost_base=state.pool_cost_base,
-                state_version=state.state_version,
-            ),
-            representative_transaction=(
-                to_booked_transaction(TransactionEvent.model_validate(representative_transaction))
-                if representative_transaction is not None
-                else None
-            ),
+        return await self._average_cost_pools.get_average_cost_pool_checkpoint_record(
+            portfolio_id=portfolio_id,
+            security_id=security_id,
         )
 
     async def upsert_average_cost_pool_checkpoint(
         self,
         checkpoint: AverageCostPoolCheckpoint,
     ) -> None:
-        payload = {
-            "portfolio_id": normalize_lookup_identifier(checkpoint.portfolio_id),
-            "security_id": normalize_lookup_identifier(checkpoint.security_id),
-            "instrument_id": normalize_lookup_identifier(checkpoint.instrument_id),
-            "representative_source_transaction_id": (
-                normalize_lookup_identifier(checkpoint.representative_source_transaction_id)
-                if checkpoint.representative_source_transaction_id
-                else None
-            ),
-            "pool_quantity": checkpoint.quantity,
-            "pool_cost_local": checkpoint.cost_local,
-            "pool_cost_base": checkpoint.cost_base,
-            "state_version": checkpoint.state_version,
-        }
-        stmt = pg_insert(AverageCostPoolState).values(**payload)
-        await self.db.execute(
-            stmt.on_conflict_do_update(
-                index_elements=["portfolio_id", "security_id"],
-                set_={
-                    field_name: getattr(stmt.excluded, field_name)
-                    for field_name in payload
-                    if field_name not in {"portfolio_id", "security_id"}
-                }
-                | {"updated_at": func.now()},
-            )
-        )
+        await self._average_cost_pools.upsert_average_cost_pool_checkpoint(checkpoint)
 
     async def apply_average_cost_pool_transition(
         self,
         transition: AverageCostPoolTransition,
     ) -> None:
-        await self._scale_existing_average_cost_sources(transition)
-        if transition.explicit_sources_after:
-            await self.update_selected_open_lot_states(
-                portfolio_id=transition.before.portfolio_id,
-                security_id=transition.before.security_id,
-                states_by_source_transaction_id=dict(transition.explicit_sources_after),
-            )
-        await self.upsert_average_cost_pool_checkpoint(transition.after)
+        await self._average_cost_pools.apply_average_cost_pool_transition(transition)
 
     async def apply_average_cost_pool_rebuild(
         self,
         plan: AverageCostPoolRebuildPlan,
     ) -> None:
-        checkpoint = plan.checkpoint
-        normalized_portfolio_id = normalize_lookup_identifier(checkpoint.portfolio_id)
-        normalized_security_id = normalize_lookup_identifier(checkpoint.security_id)
-        await self.db.execute(
-            update(PositionLotState)
-            .where(
-                func.trim(PositionLotState.portfolio_id) == normalized_portfolio_id,
-                func.trim(PositionLotState.security_id) == normalized_security_id,
-            )
-            .values(
-                open_quantity=Decimal(0),
-                lot_cost_local=Decimal(0),
-                lot_cost_base=Decimal(0),
-                updated_at=func.now(),
-            )
-        )
-
-        payloads = []
-        for source_transaction in plan.source_transactions:
-            payload = buy_lot_state_payload(source_transaction)
-            state = plan.source_states.get(source_transaction.transaction_id)
-            payload.update(
-                open_quantity=state.quantity if state is not None else Decimal(0),
-                lot_cost_local=state.cost_local if state is not None else Decimal(0),
-                lot_cost_base=state.cost_base if state is not None else Decimal(0),
-            )
-            payloads.append(payload)
-        for offset in range(0, len(payloads), self.AVERAGE_COST_REBUILD_UPSERT_CHUNK_SIZE):
-            stmt = pg_insert(PositionLotState).values(
-                payloads[offset : offset + self.AVERAGE_COST_REBUILD_UPSERT_CHUNK_SIZE]
-            )
-            await self.db.execute(
-                stmt.on_conflict_do_update(
-                    index_elements=["source_transaction_id"],
-                    set_=mutable_lot_state_fields(stmt),
-                )
-            )
-
-        await self.upsert_average_cost_pool_checkpoint(checkpoint)
+        await self._average_cost_pools.apply_average_cost_pool_rebuild(plan)
 
     async def get_average_cost_pool_persisted_summary(
         self,
@@ -354,144 +236,10 @@ class CostCalculatorRepository:
         portfolio_id: str,
         security_id: str,
     ) -> AverageCostPoolPersistedSummary:
-        normalized_portfolio_id = normalize_lookup_identifier(portfolio_id)
-        normalized_security_id = normalize_lookup_identifier(security_id)
-        pool = (
-            (
-                await self.db.execute(
-                    select(AverageCostPoolState).where(
-                        AverageCostPoolState.portfolio_id == normalized_portfolio_id,
-                        AverageCostPoolState.security_id == normalized_security_id,
-                    )
-                )
-            )
-            .scalars()
-            .first()
+        return await self._average_cost_pools.get_average_cost_pool_persisted_summary(
+            portfolio_id=portfolio_id,
+            security_id=security_id,
         )
-        source_count, source_quantity, source_cost_local, source_cost_base = (
-            await self.db.execute(
-                select(
-                    func.count(PositionLotState.id),
-                    func.coalesce(func.sum(PositionLotState.open_quantity), Decimal(0)),
-                    func.coalesce(func.sum(PositionLotState.lot_cost_local), Decimal(0)),
-                    func.coalesce(func.sum(PositionLotState.lot_cost_base), Decimal(0)),
-                ).where(
-                    func.trim(PositionLotState.portfolio_id) == normalized_portfolio_id,
-                    func.trim(PositionLotState.security_id) == normalized_security_id,
-                )
-            )
-        ).one()
-        return AverageCostPoolPersistedSummary(
-            source_count=int(source_count),
-            source_quantity=source_quantity,
-            source_cost_local=source_cost_local,
-            source_cost_base=source_cost_base,
-            pool_quantity=pool.pool_quantity if pool is not None else None,
-            pool_cost_local=pool.pool_cost_local if pool is not None else None,
-            pool_cost_base=pool.pool_cost_base if pool is not None else None,
-        )
-
-    async def _scale_existing_average_cost_sources(
-        self,
-        transition: AverageCostPoolTransition,
-    ) -> None:
-        before = transition.before
-        after = transition.existing_sources_after
-        if before.quantity == Decimal(0) or after == before.as_open_lot_state():
-            return
-
-        predicates = [
-            func.trim(PositionLotState.portfolio_id)
-            == normalize_lookup_identifier(before.portfolio_id),
-            func.trim(PositionLotState.security_id)
-            == normalize_lookup_identifier(before.security_id),
-        ]
-        explicit_source_ids = set(transition.explicit_sources_after)
-        if explicit_source_ids:
-            predicates.append(PositionLotState.source_transaction_id.not_in(explicit_source_ids))
-
-        if after.quantity == Decimal(0):
-            result = await self.db.execute(
-                update(PositionLotState)
-                .where(*predicates)
-                .values(
-                    open_quantity=Decimal(0),
-                    lot_cost_local=Decimal(0),
-                    lot_cost_base=Decimal(0),
-                    updated_at=func.now(),
-                )
-            )
-            if result.rowcount < 1:
-                raise ValueError("Average cost pool close found no persisted source lots")
-            return
-
-        representative_source_id = before.representative_source_transaction_id
-        if representative_source_id is None:
-            raise ValueError("Open average cost pool has no representative source")
-        non_residual_predicates = [
-            *predicates,
-            PositionLotState.source_transaction_id != representative_source_id,
-        ]
-        await self.db.execute(
-            update(PositionLotState)
-            .where(*non_residual_predicates)
-            .values(
-                open_quantity=_scaled_persisted_value(
-                    PositionLotState.open_quantity,
-                    before=before.quantity,
-                    after=after.quantity,
-                    round_down=True,
-                ),
-                lot_cost_local=_scaled_persisted_value(
-                    PositionLotState.lot_cost_local,
-                    before=before.cost_local,
-                    after=after.cost_local,
-                    round_down=False,
-                ),
-                lot_cost_base=_scaled_persisted_value(
-                    PositionLotState.lot_cost_base,
-                    before=before.cost_base,
-                    after=after.cost_base,
-                    round_down=False,
-                ),
-                updated_at=func.now(),
-            )
-        )
-        allocated_quantity, allocated_cost_local, allocated_cost_base = (
-            await self.db.execute(
-                select(
-                    func.coalesce(func.sum(PositionLotState.open_quantity), Decimal(0)),
-                    func.coalesce(func.sum(PositionLotState.lot_cost_local), Decimal(0)),
-                    func.coalesce(func.sum(PositionLotState.lot_cost_base), Decimal(0)),
-                ).where(*non_residual_predicates)
-            )
-        ).one()
-        residual_state = OpenLotState(
-            quantity=after.quantity - allocated_quantity,
-            cost_local=after.cost_local - allocated_cost_local,
-            cost_base=after.cost_base - allocated_cost_base,
-        )
-        if (
-            residual_state.quantity < Decimal(0)
-            or residual_state.cost_local < Decimal(0)
-            or residual_state.cost_base < Decimal(0)
-        ):
-            raise ValueError("Average cost source allocation exceeds the target pool aggregate")
-        residual_result = await self.db.execute(
-            update(PositionLotState)
-            .where(
-                *predicates,
-                PositionLotState.source_transaction_id == representative_source_id,
-            )
-            .values(
-                open_quantity=residual_state.quantity,
-                lot_cost_local=residual_state.cost_local,
-                lot_cost_base=residual_state.cost_base,
-                updated_at=func.now(),
-            )
-        )
-        if residual_result.rowcount != 1:
-            raise ValueError("Average cost pool representative source was not updated exactly once")
 
     async def get_open_lot_checkpoint_records(
         self, *, portfolio_id: str, security_id: str
@@ -700,29 +448,11 @@ class CostCalculatorRepository:
         states_by_source_transaction_id: dict[str, OpenLotState],
     ) -> None:
         """Update an explicitly selected lot subset without closing omitted open lots."""
-        if not states_by_source_transaction_id:
-            return
-
-        normalized_portfolio_id = normalize_lookup_identifier(portfolio_id)
-        normalized_security_id = normalize_lookup_identifier(security_id)
-        source_transaction_ids = set(states_by_source_transaction_id)
-        stmt = select(PositionLotState).where(
-            func.trim(PositionLotState.portfolio_id) == normalized_portfolio_id,
-            func.trim(PositionLotState.security_id) == normalized_security_id,
-            PositionLotState.source_transaction_id.in_(source_transaction_ids),
+        await self._average_cost_pools.update_selected_open_lot_states(
+            portfolio_id=portfolio_id,
+            security_id=security_id,
+            states_by_source_transaction_id=states_by_source_transaction_id,
         )
-        lot_rows = (await self.db.execute(stmt)).scalars().all()
-        persisted_source_ids = {lot_row.source_transaction_id for lot_row in lot_rows}
-        missing_source_ids = source_transaction_ids - persisted_source_ids
-        if missing_source_ids:
-            missing_ids = ", ".join(sorted(missing_source_ids))
-            raise ValueError(f"Selected cost-basis source lots are missing: {missing_ids}")
-
-        for lot_row in lot_rows:
-            state = states_by_source_transaction_id[lot_row.source_transaction_id]
-            lot_row.open_quantity = state.quantity
-            lot_row.lot_cost_local = state.cost_local
-            lot_row.lot_cost_base = state.cost_base
 
     async def upsert_accrued_income_offset_state(
         self, transaction_result: EngineTransaction
